Gui stuff.
Added R-tree for better GUI widget lookup. 
Disabled regular GUI int syncing.
Added mouseOver() for widgets.
Rewrote widgets to new R-tree system.
Added heat info to EBF.
Added loss to GT cable screen.
Improve info rendering.

+++ src/main/java/muramasa/antimatter/gui/widget/AbstractSwitchWidget.java
@@ -49,7 +49,7 @@ public abstract class AbstractSwitchWidget extends ButtonWidget {
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void render(MatrixStack matrixStack, double mouseX, double mouseY, float partialTicks) {
         super.render(matrixStack, mouseX, mouseY, partialTicks);
         Minecraft minecraft = Minecraft.getInstance();
         minecraft.getTextureManager().bindTexture(res);
