+++ src/main/java/muramasa/antimatter/gui/widget/TextWidget.java
@@ -24,7 +24,7 @@ public class TextWidget extends Widget {
     }
 
     @Override
-    public void render(MatrixStack matrixStack, int mouseX, int mouseY, float partialTicks) {
+    public void render(MatrixStack matrixStack, double mouseX, double mouseY, float partialTicks) {
         this.drawText(matrixStack, new StringTextComponent(getter.apply(this)), realX(), realY(), color);
     }
 }
